renamed and reorganized module imports for better structure
