+++ core/config.json
